docker_build/requirements.txt
python-docx
requests
bottle